+++ src/MapGen/Models/Terrain.hs
@@ -4,7 +4,7 @@ module MapGen.Models.Terrain (
 
 import System.Random (Random, random, randomR)
 
-data Terrain = Plains | Shallows | Ocean | Forest | Hills | Mountains | Peaks deriving (Show, Eq, Ord, Enum, Bounded)
+data Terrain = Aerie | Plains | Shallows | Ocean | Forest | Hills | Mountains | Peaks deriving (Show, Eq, Ord, Enum, Read, Bounded)
 
 instance Random Terrain where
   randomR (lo, hi) g =